アンカーリンクのコンポーネントにpタグを含めないように変更
