+++ src/components/contents/shared/TextUrlToLink.tsx
@@ -1,15 +1,13 @@
 import React, { VFC } from 'react'
-import styled from 'styled-components'
-import { Text } from 'smarthr-ui'
 import reactStringReplace from 'react-string-replace'
 
 type Props = {
   text: string
 }
 
-export const TextAnchor: VFC<Props> = ({ text }) => {
+export const TextUrlToLink: VFC<Props> = ({ text }) => {
   return (
-    <StyledText as="p">
+    <>
       {reactStringReplace(text, /(https?:\/\/\S+)/g, (match: string, strIndex: number) => {
         return (
           <a key={strIndex} href={match}>
@@ -17,11 +15,6 @@ export const TextAnchor: VFC<Props> = ({ text }) => {
           </a>
         )
       })}
-    </StyledText>
+    </>
   )
 }
-
-const StyledText = styled(Text)`
-  white-space: pre-wrap;
-  word-wrap: break-word;
-`
